+++ example/android/app/src/main/kotlin/com/amap/location/amap_location_flutter_plugin_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.amap.location.amap_location_flutter_plugin_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()